DDD | Create ApplicationLayerAssertkExtensions.kt
