+++ backend/bialydunajec-camp-edition/application/src/main/kotlin/org/bialydunajec/campedition/application/query/api/CampEditionQueryGateway.kt
@@ -1,11 +1,12 @@
 package org.bialydunajec.campedition.application.query.api
 
 import org.bialydunajec.campedition.application.query.readmodel.CampEditionDomainModelReader
+import org.bialydunajec.ddd.application.base.query.QueryGateway
 import org.springframework.stereotype.Component
 
-class CampEditionQueryGateway internal constructor(private val domainModelReader: CampEditionDomainModelReader) {
+class CampEditionQueryGateway internal constructor(private val domainModelReader: CampEditionDomainModelReader) : QueryGateway<CampEditionQuery> {
 
-    fun process(query: CampEditionQuery) = when (query) {
+    override fun process(query: CampEditionQuery) = when (query) {
         is CampEditionQuery.All -> process(query)
         is CampEditionQuery.ById -> process(query)
     }
