+++ backend/bialydunajec-ddd/application/src/main/kotlin/org/bialydunajec/ddd/application/base/query/QueryGateway.kt
@@ -1,3 +1,7 @@
 package org.bialydunajec.ddd.application.base.query
 
-interface QueryGateway
+interface QueryGateway<QueryType: Query>{
+
+    fun process(query: QueryType): Any?
+}
+
